# packages/backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Montessori ERP & LMS API
  description: |
    Multi-tenant, role-based ERP & Learning Management System for Montessori schools.
    
    ## Authentication
    All routes (except `/api/v1/auth/*`) require a JWT Bearer token in the `Authorization` header.
    
    ## Roles
    - **admin**: Full access to all modules
    - **teacher**: Students, attendance, observations, curriculum, messages
    - **parent**: Read-only access to own children, fees, messages
    - **student**: Read-only curriculum and announcements
  version: 1.0.0
  contact:
    name: Montessori ERP Support
servers:
  - url: http://localhost:4001
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
...
paths:
...
  /api/v1/curriculum/tree:
    get:
      tags: [Curriculum]
      summary: Full curriculum tree with progress overlay
      security:
        - bearerAuth: []
      parameters:
        - name: studentId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Curriculum tree with progress
  /api/v1/curriculum/areas:
    get:
      tags: [Curriculum]
      summary: List curriculum areas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Curriculum areas list
  /api/v1/curriculum/items:
    get:
      tags: [Curriculum]
      summary: List curriculum items
      security:
        - bearerAuth: []
      parameters:
        - name: areaId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Curriculum items list
  /api/v1/curriculum/lesson-plans:
    get:
      tags: [Curriculum]
      summary: List lesson plans
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lesson plans list
    post:
      tags: [Curriculum]
      summary: Create lesson plan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, date]
              properties:
                title:
                  type: string
                curriculumItemId:
                  type: string
                content:
                  type: string
                date:
                  type: string
                  format: date
                classId:
                  type: string
      responses:
        '201':
          description: Lesson plan created
  /api/v1/curriculum/lesson-plans/{id}:
    get:
      tags: [Curriculum]
      summary: Get lesson plan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lesson plan details
    patch:
      tags: [Curriculum]
      summary: Update lesson plan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lesson plan updated
    delete:
      tags: [Curriculum]
      summary: Delete lesson plan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Lesson plan deleted
  /api/v1/curriculum/progress/student/{studentId}:
    get:
      tags: [Curriculum]
      summary: Get student's curriculum progress
      security:
        - bearerAuth: []
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student progress

  # ==================== FINANCE ====================
...
  /api/v1/communication/messages:
    get:
      tags: [Communication]
      summary: List messages
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Messages list
    post:
      tags: [Communication]
      summary: Send message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content, recipientIds]
              properties:
                content:
                  type: string
                recipientIds:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Message sent
  /api/v1/communication/messages/{messageId}/read:
    patch:
      tags: [Communication]
      summary: Mark message as read
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Message marked as read
  /api/v1/communication/announcements:
    get:
      tags: [Communication]
      summary: List announcements
      security:
        - bearerAuth: []
      parameters:
        - name: classId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Announcements list
    post:
      tags: [Communication]
      summary: Create announcement (admin/teacher)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, body]
              properties:
                title:
                  type: string
                body:
                  type: string
                classId:
                  type: string
      responses:
        '201':
          description: Announcement created
  /api/v1/communication/announcements/{announcementId}/read:
    patch:
      tags: [Communication]
      summary: Mark announcement as read
      security:
        - bearerAuth: []
      parameters:
        - name: announcementId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Announcement marked as read
  /api/v1/communication/unread-counts:
    get:
      tags: [Communication]
      summary: Get unread message/announcement counts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Unread counts

  # ==================== GAMIFICATION ====================
...